+++ src/block/icon-list/deprecated-files/icons.js
